Fix script bugs, attempt to clean up actor update logic to fix rolling bug.
